Fix django-cms redirect middleware in django-cms 3.0, albiet badly.
Django-cms 3 now uses TemplateResponse objects to render templates (instead of plain old render_to_response in 2.X), and for some reason, any exceptions thrown inside TemplateResponses aren't caught by the process_exception method in middleware. So, load an extra attribute on to the request object (formDesignerRedirect) in the plugin renderer and use a process_response middleware method to detect it's presence, and return the HttpRedirect in such case.